// frontend/src/app/mgf/core/models/request-cas-param.ts
export interface RequestCasParam {
  dateDebut?: string;
  dateFin?: string;
  idTypeCas?: number;
  telephone?: string;
  idRegion?: number;
  idPrefecture?: number;
  idCommune?: number;
  idQuartier?: number;
  idSecteur?: number;
  ageDebut?: number;
  ageFin?: number;
  role?: string;
  page?: number; 
  size?: number;
}